Use Either in CachingState
This allows us to expose the cache key and the captured execution state as regular values when caching is enabled (and thus we must know them), and as Optional's when caching is disabled and we might or might not know these. Doing so allows simplifications in BuildCacheStep, and perhaps other places.